style: Reorganizando pacotes de testes unitários
