+++ src/test/java/br/com/devquest/api/unittests/mocks/MockTokenDTO.java
@@ -1,4 +1,4 @@
-package br.com.devquest.api.unittests.security.mocks;
+package br.com.devquest.api.unittests.mocks;
 
 import br.com.devquest.api.model.dtos.security.TokenDTO;
 import lombok.NoArgsConstructor;
